Move calculations of income and next period wealth to egm step.
