+++ src/consumption_retirement_model.py
@@ -74,123 +74,6 @@ def inverse_marginal_utility_crra(
     return inverse_marginal_utility
 
 
-def compute_current_income(
-    period: int, shock: float, params: pd.DataFrame, options
-) -> float:
-    """Computes the current level of deterministic and stochastic income.
-
-    Note that income is paid at the end of the current period, i.e. after
-    the (potential) labor supply choice has been made. This is equivalent to
-    allowing income to be dependent on a lagged choice of labor supply.
-
-    The agent starts working in period t = 0.
-    Relevant for the wage equation (deterministic income) are age-dependent
-    coefficients of work experience:
-
-    labor_income = constant + alpha_1 * age + alpha_2 * age**2
-
-    They include a constant as well as two coefficents on age and age squared,
-    respectively. Note that the last one (alpha_2) typically has a negative sign.
-
-    Args:
-        period (int): Curent period t.
-        shock (float): Stochastic shock on labor income, which may or may not
-            be normally distributed.
-        params (pd.DataFrame): Model parameters indexed with multi-index of the
-            form ("category", "name") and two columns ["value", "comment"].
-            Relevant here are the coefficients of the wage equation.
-        options (dict): Options dictionary.
-
-    Returns:
-        stochastic_income (float): End of period income composed of a
-            deterministic component, i.e. age-dependent labor income, and a
-            stochastic shock.
-    """
-    # For simplicity, assume current_age - min_age = experience
-    # TODO: Allow age and work experience to differ,
-    # i.e. allow for unemployment spells
-    min_age = options["min_age"]
-    age = period + min_age
-
-    # Determinisctic component of income depending on experience
-    # labor_income = constant + alpha_1 * age + alpha_2 * age**2
-    exp_coeffs = np.asarray(params.loc["wage", "value"])
-    labor_income = exp_coeffs @ (age ** np.arange(len(exp_coeffs)))
-
-    stochastic_income = np.exp(labor_income + shock)
-
-    return stochastic_income
-
-
-def compute_next_period_wealth_matrix(
-    period: int,
-    state: int,
-    savings: float,
-    quad_points: float,
-    params: pd.DataFrame,
-    options: Dict[str, int],
-) -> float:
-    """Computes all possible levels of next period wealth M_(t+1)
-
-    Args:
-        period (int): Current period t.
-        state (int): State of the agent, e.g. 0 = "retirement", 1 = "working".
-        savings_grid (np.ndarray): Array of shape n_grid_wealth denoting the
-            exogenous savings grid.
-        quad_points (np.ndarray): Array of shape (n_quad_stochastic,)
-            containing (normally distributed) stochastic components.
-        params (pd.DataFrame): Model parameters indexed with multi-index of the
-            form ("category", "name") and two columns ["value", "comment"].
-        options (dict): Options dictionary.
-
-    Returns:
-        next_period_wealth (np.ndarray): Array of all possible next period
-            wealths with shape (n_quad_stochastic, n_grid_wealth).
-    """
-    r = params.loc[("assets", "interest_rate"), "value"]
-    sigma = params.loc[("shocks", "sigma"), "value"]
-
-    n_grid_wealth = options["grid_points_wealth"]
-    n_quad_stochastic = options["quadrature_points_stochastic"]
-
-    shocks = quad_points * sigma
-    next_period_income = compute_current_income(period + 1, shocks, params, options)
-
-    next_period_wealth = np.full(
-        (n_grid_wealth, n_quad_stochastic), next_period_income * state,
-    ).T + np.full((n_quad_stochastic, n_grid_wealth), savings * (1 + r))
-
-    # Retirement safety net, only in retirement model with 2 discrete choices
-    if options["n_discrete_choices"] > 1:
-        consumption_floor = params.loc[("assets", "consumption_floor"), "value"]
-        next_period_wealth[next_period_wealth < consumption_floor] = consumption_floor
-
-    return next_period_wealth
-
-
-def compute_marginal_wealth_matrix(
-    params: pd.DataFrame, options: Dict[str, int]
-) -> np.ndarray:
-    """Computes marginal next period wealth.
-
-    Args:
-        params (pd.DataFrame): Model parameters indexed with multi-index of the
-            form ("category", "name") and two columns ["value", "comment"].
-        options (dict): Options dictionary.
-
-    Returns:
-        marginal_wealth (np.ndarray):  Array of all possible next period
-            marginal wealths. Shape (n_quad_stochastic, n_grid_wealth).
-    """
-    n_grid_wealth = options["grid_points_wealth"]
-    n_quad_stochastic = options["quadrature_points_stochastic"]
-    r = params.loc[("assets", "interest_rate"), "value"]
-
-    marginal_wealth = np.full((n_quad_stochastic, n_grid_wealth), (1 + r))
-
-    return marginal_wealth
-
-
 def compute_value_function(
     value: List[np.ndarray],
     next_period_wealth_matrix: np.ndarray,
